+++ src/ploomber/spec/DAGSpec.py
@@ -57,6 +57,31 @@ interface, we don't call it config because there is a DAGConfig object in
 the Python API and this might cause confusion
 
 All other sections should represent valid DAG properties.
+
+
+Partials [Provisional name]
+---------------------------
+Partials were introduced to support a common ML use case: train and serving
+pipelines. If we describe a training pipeline as a DAG, the only difference
+with its serving counterpart are the tasks (can be more than one) at the
+beginning (train: get historical data, serve: get single data point) and the
+end (train: fit a model, serve: load a model and predict). Everything that
+happens in the middle should be the same to avoid training-serving skew. There
+are two specific use cases: batch and online.
+
+To define a batch serving pipeline we define three files: pipeline.yaml
+(train DAG), pipeline-serve (serve DAG), and pipeline-features.yaml (partial).
+The first two follow the DAGSpec schema and use
+``import_tasks_from: pipeline-features.yaml``. Training is done via
+``ploomber build --entry-point pipeline.yaml`` and serving via
+``ploomber build --entry-point pipeline-serve.yaml``.
+
+An online API has a different mechanism because the inference pipeline needs
+a Python API to interface with a web framework, rpc, or similar. This logic
+is implemented by OnlineDAG, which takes a partial definition
+(e.g. ``pipeline-features.yaml``) and builds an in-memory DAG that can make
+predictions using ``OnlineDAG().predict()``. See ``OnlineDAG`` documentation
+for details.
 """
 import os
 import yaml
@@ -67,12 +92,14 @@ from glob import iglob
 from itertools import chain
 import pprint
 
-from ploomber import products
-from ploomber import DAG, tasks, SourceLoader
+from ploomber import products, tasks
+from ploomber.dag.DAG import DAG
+from ploomber.placeholders.SourceLoader import SourceLoader
 from ploomber.util.util import (load_dotted_path, find_file_recursively,
                                 call_with_dictionary, add_to_sys_path)
 from ploomber.spec.TaskSpec import TaskSpec, suffix2taskclass
 from ploomber.util import validate
+from ploomber.util import default
 from ploomber.dag.DAGConfiguration import DAGConfiguration
 from ploomber.exceptions import DAGSpecInitializationError
 from ploomber.env.EnvDict import EnvDict
@@ -116,30 +143,60 @@ class DAGSpec(MutableMapping):
         with the dotted path, which means some verifications such as import
         statements in that function's module are delayed until the pipeline
         is executed. This also applies to placeholders loaded using a
-        SourceLoader, if a placeholder does not exist, it will return
-        None instead of raising an error
+        SourceLoader, if a placeholder exists, it will return the path
+        to it, instead of an initialized Placeholder object, if it doesn't,
+        it will return None instead of raising an error.
 
     reload : bool, optional
         Reloads modules before getting dotted paths. Has no effect if
         lazy_import=True
     """
-    def __init__(self, data, env=None, lazy_import=False, reload=False):
+
+    # NOTE: lazy_import is used where we need to initialized a a spec but don't
+    # plan on running it. One use case is when exporting to Argo or Airflow:
+    # we don't want to raise errors if some dependency is missing because
+    # it can happen that the environment exporting the dag does not have
+    # all the dependencies required to run it. The second use case is when
+    # running "ploomber scaffold", we want to use DAGSpec machinery to parse
+    # the yaml spec and the use such information to add the task in the
+    # appropriate place, this by construction, means that the spec as it is
+    # cannot be converted to a dag yet, since it has at least one task
+    # whose source does not exist
+    def __init__(self,
+                 data,
+                 env=None,
+                 lazy_import=False,
+                 reload=False,
+                 parent_path=None):
         if isinstance(data, (str, Path)):
-            path = data
+            if parent_path is not None:
+                raise ValueError('parent_path must be None when '
+                                 f'initializing {type(self).__name__} with '
+                                 'a path to a YAML spec')
+            # this is only used to display an error message with the path
+            # to the loaded file
+            path_for_errors = data
             # resolve the parent path to make sources and products unambiguous
             # even if the current working directory changes
-            self._parent_path = str(Path(data).parent.resolve())
+            path_to_entry_point = Path(data).resolve()
+            self._parent_path = str(path_to_entry_point.parent)
 
             with open(str(data)) as f:
                 data = yaml.load(f, Loader=yaml.SafeLoader)
 
-            env_path = Path(self._parent_path, 'env.yaml')
-
-            if env is None and env_path.exists():
-                env = str(env_path)
         else:
-            path = None
-            self._parent_path = None
+            path_for_errors = None
+            # FIXME: add test cases, some of those features wont work if
+            # _parent_path is None. We should make sure that we either raise
+            # an error if _parent_path is needed or use the current working
+            # directory if it's appropriate - this is mostly to make relative
+            # paths consistent: they should be relative to the file that
+            # contains them
+            self._parent_path = (None if not parent_path else str(
+                Path(parent_path).resolve()))
+
+        if env is None and self._parent_path is not None:
+            env = default.path_to_env(self._parent_path)
 
         self.data = data
 
@@ -147,7 +204,7 @@ class DAGSpec(MutableMapping):
             self.data = {'tasks': self.data}
 
         # validate keys defined at the top (nested keys are not validated here)
-        self._validate_top_keys(self.data, path)
+        self._validate_top_keys(self.data, path_for_errors)
 
         logger.debug('DAGSpec enviroment:\n%s', pp.pformat(env))
 
@@ -170,13 +227,35 @@ class DAGSpec(MutableMapping):
         if 'location' not in self.data:
 
             Meta.initialize_inplace(self.data)
+
             import_tasks_from = self.data['meta']['import_tasks_from']
 
             if import_tasks_from is not None:
-                imported = yaml.safe_load(Path(import_tasks_from).read_text())
+                # when using a relative path in "import_tasks_from", we must
+                # make it absolute...
+                if not Path(import_tasks_from).is_absolute():
+                    # use _parent_path if there is one
+                    if self._parent_path:
+                        self.data['meta']['import_tasks_from'] = str(
+                            Path(self._parent_path, import_tasks_from))
+                    # otherwise just make it absolute
+                    else:
+                        self.data['meta']['import_tasks_from'] = str(
+                            Path(import_tasks_from).resolve())
+
+                imported = yaml.safe_load(
+                    Path(self.data['meta']['import_tasks_from']).read_text())
 
                 if self.env is not None:
                     imported = expand_raw_dictionaries(imported, self.env)
+
+                # relative paths here are relative to the file where they
+                # are declared
+                base_path = Path(self.data['meta']['import_tasks_from']).parent
+
+                for task in imported:
+                    add_base_path_to_source_if_relative(task,
+                                                        base_path=base_path)
 
                 self.data['tasks'].extend(imported)
 
@@ -210,7 +289,10 @@ class DAGSpec(MutableMapping):
                 raise KeyError('If specifying dag through a "location" key '
                                'it must be the unique key in the spec')
         else:
-            valid = {'meta', 'config', 'clients', 'tasks'}
+            valid = {
+                'meta', 'config', 'clients', 'tasks', 'serializer',
+                'unserializer'
+            }
             validate.keys(valid, spec.keys(), name='dag spec')
 
     def __getitem__(self, key):
@@ -263,6 +345,12 @@ class DAGSpec(MutableMapping):
 
         if clients:
             init_clients(dag, clients)
+
+        # FIXME: this violates lazy_import, we must change DAG's implementation
+        # to accept strings as attribute and load them until they are called
+        for attr in ['serializer', 'unserializer']:
+            if attr in self:
+                setattr(dag, attr, load_dotted_path(self[attr]))
 
         process_tasks(dag, self, root_path=self._parent_path)
 
@@ -517,3 +605,13 @@ def normalize_task(task):
         return {'source': task}
     else:
         return task
+
+
+def add_base_path_to_source_if_relative(task, base_path):
+    path = Path(task['source'])
+    relative_source = not path.is_absolute()
+
+    # must be a relative source with a valid extension, otherwise, it can
+    # be a dotted path
+    if relative_source and path.suffix in set(suffix2taskclass):
+        task['source'] = str(Path(base_path, task['source']).resolve())