+++ src/ploomber/tasks/tasks.py
@@ -25,7 +25,7 @@ def _unserialize_params(params_original, unserializer):
     params = params_original.to_dict()
 
     params['upstream'] = {
-        k: unserializer(v)
+        k: unserializer(product=v)
         for k, v in params['upstream'].items()
     }
 
@@ -74,10 +74,10 @@ class PythonCallable(Task):
                  params=None,
                  unserializer=None,
                  serializer=None):
-        kwargs = dict(hot_reload=dag._params.hot_reload,
-                      needs_product=serializer is None)
-        self._source = type(self)._init_source(source, kwargs)
         self._serializer = serializer or dag.serializer
+        kwargs = dict(hot_reload=dag._params.hot_reload,
+                      needs_product=self._serializer is None)
+        self._source = type(self)._init_source(source, kwargs)
         self._unserializer = unserializer or dag.unserializer
         super().__init__(product, dag, name, params)