+++ tests/cli/test_customparser.py
@@ -1,6 +1,11 @@
+from pathlib import Path
 import sys
+from unittest.mock import Mock
+
 import pytest
-from ploomber.cli.parsers import CustomParser
+
+from ploomber.cli.parsers import CustomParser, _custom_command
+from ploomber.cli import parsers
 
 
 def test_custom_parser_static_args():
@@ -50,3 +55,53 @@ def test_default_loaded_from_env_var(monkeypatch):
 
     args = parser.parse_args()
     assert args.entry_point == 'dag.yaml'
+
+
+def test_dagspec_initialization_from_yaml(tmp_nbs_nested, monkeypatch):
+    """
+    DAGSpec can be initialized with a path to a spec or a dictionary, but
+    they have a slightly different behavior. This checks that we initialize
+    with the path
+    """
+    mock = Mock(wraps=parsers.DAGSpec)
+
+    monkeypatch.setattr(sys, 'argv', ['python'])
+    monkeypatch.setattr(parsers, 'DAGSpec', mock)
+
+    parser = CustomParser()
+
+    with parser:
+        pass
+
+    dag, args = _custom_command(parser)
+
+    mock.assert_called_once_with('pipeline.yaml')
+
+
+def test_dagspec_initialization_from_yaml_and_env(tmp_nbs, monkeypatch):
+    """
+    DAGSpec can be initialized with a path to a spec or a dictionary, but
+    they have a slightly different behavior. This checks that we initialize
+    with the path
+    """
+    mock_dagspec = Mock(wraps=parsers.DAGSpec)
+    mock_default_path_to_env = Mock(wraps=parsers.default.path_to_env)
+    mock_envdict = Mock(wraps=parsers.EnvDict)
+
+    monkeypatch.setattr(sys, 'argv', ['python'])
+    monkeypatch.setattr(parsers, 'DAGSpec', mock_dagspec)
+    monkeypatch.setattr(parsers.default, 'path_to_env',
+                        mock_default_path_to_env)
+    monkeypatch.setattr(parsers, 'EnvDict', mock_envdict)
+
+    parser = CustomParser()
+
+    with parser:
+        pass
+
+    dag, args = _custom_command(parser)
+
+    mock_dagspec.assert_called_once_with('pipeline.yaml',
+                                         env={'sample': False})
+    mock_default_path_to_env.assert_called_once_with(Path('.'))
+    mock_envdict.assert_called_once_with(str(Path('env.yaml').resolve()))